Update treeplot.R
I add the panel plot chunk in here and NO need to run score_compare prior to plot either.
I haven't been able to figure out how to add legend to the tree, but I left the order of the content to line 94;
When compute prevalence, we have to remove the prefix to extract the counts. I don't know how to get the prevalence w.r.t the prefix. This means the prevalence for O_H and S_H with the same diagnostic/atc code are the same.:-/
I tried facet_grid() and all the text were entangled with each other, so I use ggarrange(). But I don't know how to align the predictors via ggarrange(). It looks civilised with 2-row arrange. 3-row looks squished. So I took the liberty to do 2 2-rows by adding the probability with CIs in. 
"for_the_tree.csv" could be used for the plot part starting from line 77.